# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Roisfaozi_queue-base_internal_modules_settings_model.CreateSettingRequest:
    properties:
      key:
        maxLength: 100
        minLength: 1
        type: string
      scope_id:
        type: string
      scope_type:
        enum:
        - tenant
        - branch
        - service
        - counter
        type: string
      value:
        type: string
      value_type:
        enum:
        - string
        - number
        - boolean
        - json
        type: string
    required:
    - key
    - scope_id
    - scope_type
    - value
    type: object
  github_com_Roisfaozi_queue-base_internal_modules_settings_model.UpdateSettingRequest:
    properties:
      is_active:
        type: boolean
      value:
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.PageMetadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_item:
        type: integer
      total_page:
        type: integer
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper:
    properties:
      data: {}
      paging:
        $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.PageMetadata'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clean and modular boilerplate for Go REST APIs with RBAC,
    Audit Logs, and WebSockets.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Clean Boilerplate API
  version: "1.0"
paths:
...
  /settings:
    post:
      consumes:
      - application/json
      description: Creates tenant, branch, service, or counter scoped setting override.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Create Setting Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_settings_model.CreateSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Create setting
      tags:
      - settings
  /settings/{id}:
    delete:
      description: Deletes scoped setting override under active tenant scope.
      parameters:
      - description: Setting ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Delete setting
      tags:
      - settings
    get:
      description: Returns setting details under active tenant scope.
      parameters:
      - description: Setting ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get setting by ID
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Updates scoped setting override under active tenant scope.
      parameters:
      - description: Setting ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Update Setting Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_settings_model.UpdateSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Update setting
      tags:
      - settings
  /settings/resolve:
    get:
      description: Resolves setting value using tenant -> branch -> service -> counter
        inheritance.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Setting key
        in: query
        name: Key
        required: true
        type: string
      - description: Branch ID
        in: query
        name: BranchID
        type: string
      - description: Service ID
        in: query
        name: ServiceID
        type: string
      - description: Counter ID
        in: query
        name: CounterID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Resolve effective setting
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    description: '"Type ''Bearer '' followed by a space and the access token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"